test-slim Phase 3 (cluster 2): merge phase-named test files into module-named files
Moves verbatim, then dedupes same-behaviour pairs. Governance coverage
(posture, grants, capability enforcement, audit) reorganised, not thinned.

Renames/merges:
- test_phase8_posture_gate.py -> test_posture_gate.py (rename only)
- test_phase10_runtime.py + test_phase6_r2_fixes.py -> test_runtime.py
- test_phase10_invocations_api.py + test_phase6_vertical_slice.py
  -> test_invocations_api.py (ONE full install->advance->grant->invoke
  vertical slice, error-path tests around it; shared fixtures)
- test_phase2_capability_grammar.py + test_phase2_grants_v2.py
  -> folded into test_capabilities.py
- test_phase1_security_fixes.py -> folded into test_advice_boundary.py

Dedup notes (8 dedups, one test per distinct failure mode):
1-2. require_admin unit tests dropped — admin-403 (admin_required)
   failure mode covered at endpoint level by the admin API tests.
3. two _derive_actor_role tests merged into one (same trust-boundary
   behaviour, both derivations asserted).
4. grants-v2 v1-grant-still-resolves merged into the existing
   require_capability happy path (NULL 0015-column assertions kept).
5. grants-v2 unique-constraint test dropped — duplicate of
   test_grant_is_idempotent.
6-7. adapter ProviderKeyMissing/ProviderUpstreamError propagation
   tests dropped — same failure modes asserted end-to-end through the
   adapter by the HTTP 422/502 translation tests in
   test_invocations_api.py.
8. phase10 happy-path reconstruction-integration test dropped —
   duplicate of the vertical slice's reconstruction step (its extra
   module.grant.created assertion folded into the slice).

Cross-references updated: AUDIT_COVERAGE_MATRIX.md (also fixes stale
test_phase10_invoke_endpoint* globs/test names), contract_review
README, test_phase7_grants_api/test_prompt_runtime/dormant comments.

Net -324 lines in backend/tests. Full suite: 805 passed, 3 failed
(known macOS sandbox rlimit baseline), 11 skipped, 1 xfailed; zero
collection errors.
